Refactored price feeds
