+++ packages/financial-templates-lib/src/price-feed/CreatePriceFeed.js
@@ -4,6 +4,7 @@ const { MedianizerPriceFeed } = require("./MedianizerPriceFeed");
 const { CryptoWatchPriceFeed } = require("./CryptoWatchPriceFeed");
 const { UniswapPriceFeed } = require("./UniswapPriceFeed");
 const { BalancerPriceFeed } = require("./BalancerPriceFeed");
+const { defaultConfigs } = require("./DefaultPriceFeedConfigs");
 
 const Uniswap = require("@uma/core/build/contracts/Uniswap.json");
 const ExpiringMultiParty = require("@uma/core/build/contracts/ExpiringMultiParty.json");
@@ -239,72 +240,6 @@ function createTokenPriceFeedForEmp(logger, web3, networker, getTime, empAddress
     return createUniswapPriceFeedForEmp(logger, web3, networker, getTime, empAddress, config);
   }
 }
-
-// Default price feed configs for currently approved identifiers.
-const defaultConfigs = {
-  "ETH/BTC": {
-    type: "medianizer",
-    pair: "ethbtc",
-    lookback: 7200,
-    minTimeBetweenUpdates: 60,
-    medianizedFeeds: [
-      { type: "cryptowatch", exchange: "coinbase-pro" },
-      { type: "cryptowatch", exchange: "binance" },
-      { type: "cryptowatch", exchange: "bitstamp" }
-    ]
-  },
-  "COMP/USD": {
-    // Kovan uses the "/"
-    type: "medianizer",
-    lookback: 7200,
-    minTimeBetweenUpdates: 60,
-    medianizedFeeds: [
-      { type: "cryptowatch", exchange: "coinbase-pro", pair: "compusd" },
-      { type: "cryptowatch", exchange: "poloniex", pair: "compusdt" },
-      { type: "cryptowatch", exchange: "ftx", pair: "compusd" }
-    ]
-  },
-  COMPUSD: {
-    // Mainnet has no "/"
-    type: "medianizer",
-    lookback: 7200,
-    minTimeBetweenUpdates: 60,
-    medianizedFeeds: [
-      { type: "cryptowatch", exchange: "coinbase-pro", pair: "compusd" },
-      { type: "cryptowatch", exchange: "poloniex", pair: "compusdt" },
-      { type: "cryptowatch", exchange: "ftx", pair: "compusd" }
-    ]
-  },
-  USDETH: {
-    type: "medianizer",
-    lookback: 7200,
-    invertPrice: true,
-    minTimeBetweenUpdates: 60,
-    medianizedFeeds: [
-      { type: "cryptowatch", exchange: "coinbase-pro", pair: "ethusd" },
-      { type: "cryptowatch", exchange: "binance", pair: "ethusdt" },
-      { type: "cryptowatch", exchange: "kraken", pair: "ethusd" }
-    ]
-  },
-  USDBTC: {
-    type: "medianizer",
-    lookback: 7200,
-    invertPrice: true,
-    minTimeBetweenUpdates: 60,
-    medianizedFeeds: [
-      { type: "cryptowatch", exchange: "coinbase-pro", pair: "btcusd" },
-      { type: "cryptowatch", exchange: "binance", pair: "btcusdt" },
-      { type: "cryptowatch", exchange: "bitstamp", pair: "btcusd" }
-    ]
-  },
-  USDPERL: {
-    type: "medianizer",
-    lookback: 7200,
-    invertPrice: true,
-    minTimeBetweenUpdates: 60,
-    medianizedFeeds: [{ type: "cryptowatch", exchange: "binance", pair: "perlusdt" }]
-  }
-};
 
 /**
  * Create a reference price feed for the EMP. Note: this is the price feed that the token is tracking.
